Refactor Prometheus Operator install/upgrade component
