+++ platform-operator/controllers/verrazzano/component/prometheus/operator/promoperator_component.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package prometheusoper
+package operator
 
 import (
 	"path/filepath"
