+++ platform-operator/controllers/verrazzano/component/registry/registry.go
@@ -15,7 +15,7 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/mysql"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/nginx"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/oam"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/prometheusoper"
+	promoperator "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/prometheus/operator"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/rancher"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/verrazzano"
@@ -63,7 +63,7 @@ func getComponents() []spi.Component {
 			mysql.NewComponent(),
 			keycloak.NewComponent(),
 			kiali.NewComponent(),
-			prometheusoper.NewComponent(),
+			promoperator.NewComponent(),
 		}
 	}
 	return componentsRegistry
